pyproject.toml
[project]
name = "baguetter"
version = "0.1.0"
description = "Baguetter is a flexible and efficient search engine library implemented in Python. It supports sparse (traditional), dense (semantic), and hybrid retrieval methods."
authors = [
    { name = "Mixedbread", email = "support@mixedbread.ai" },
    { name = "Julius Lipp", email = "julius@mixedbread.ai" },
    { name = "Sean Lee", email = "sean@mixedbread.ai" },
]
dependencies = [
    "autofaiss>=2.17.0",
    "faiss-cpu>=1.8.0.post1",
    "krovetzstemmer>=0.8",
    "pystemmer>=2.2.0.1",
    "nltk>=3.8.1",
    "numba>=0.60.0",
    "numpy>=1.26.4",
    "optuna>=3.6.1",
    "ranx>=0.3.20",
    "scikit-learn>=1.5.1",
    "tqdm>=4.66.4",
    "unidecode>=1.3.8",
    "usearch>=2.12.0",
    "python-dotenv>=1.0.1",
    "beir>=2.0.0",
    "openpyxl>=3.1.5",
    "datasets>=2.20.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { file = "LICENSE" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ipykernel>=6.29.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
    "uv>=0.2.30",
    "jupyter>=1.0.0",
    "notebook>=6.5.4",
]

[tool.setuptools.packages.find]
include = ["baguetter", "baguetter.*"]

[tool.setuptools.package-data]
baguetter = ["py.typed"]
